+++ src/compas_ghpython/scene/capsuleobject.py
@@ -9,19 +9,7 @@ from .sceneobject import GHSceneObject
 
 
 class CapsuleObject(GHSceneObject, GeometryObject):
-    """Scene object for drawing capsule shapes.
-
-    Parameters
-    ----------
-    capsule : :class:`compas.geometry.Capsule`
-        A COMPAS capsule.
-    **kwargs : dict, optional
-        Additional keyword arguments.
-
-    """
-
-    def __init__(self, capsule, **kwargs):
-        super(CapsuleObject, self).__init__(geometry=capsule, **kwargs)
+    """Scene object for drawing capsule shapes."""
 
     def draw(self):
         """Draw the capsule associated with the scene object.