+++ src/compas_ghpython/scene/ellipseobject.py
@@ -9,19 +9,7 @@ from .sceneobject import GHSceneObject
 
 
 class EllipseObject(GHSceneObject, GeometryObject):
-    """Scene object for drawing ellipses.
-
-    Parameters
-    ----------
-    ellipse : :class:`compas.geometry.Ellipse`
-        A COMPAS ellipse.
-    **kwargs : dict, optional
-        Additional keyword arguments.
-
-    """
-
-    def __init__(self, ellipse, **kwargs):
-        super(EllipseObject, self).__init__(geometry=ellipse, **kwargs)
+    """Scene object for drawing ellipses."""
 
     def draw(self):
         """Draw the ellipse.