+++ src/compas_ghpython/scene/sphereobject.py
@@ -9,19 +9,7 @@ from .sceneobject import GHSceneObject
 
 
 class SphereObject(GHSceneObject, GeometryObject):
-    """Scene object for drawing sphere shapes.
-
-    Parameters
-    ----------
-    sphere : :class:`compas.geometry.Sphere`
-        A COMPAS sphere.
-    **kwargs : dict, optional
-        Additional keyword arguments.
-
-    """
-
-    def __init__(self, sphere, **kwargs):
-        super(SphereObject, self).__init__(geometry=sphere, **kwargs)
+    """Scene object for drawing sphere shapes."""
 
     def draw(self):
         """Draw the sphere associated with the scene object.