+++ src/compas_rhino/scene/capsuleobject.py
@@ -12,19 +12,7 @@ from .sceneobject import RhinoSceneObject
 
 
 class RhinoCapsuleObject(RhinoSceneObject, GeometryObject):
-    """Scene object for drawing capsule shapes.
-
-    Parameters
-    ----------
-    capsule : :class:`compas.geometry.Capsule`
-        A COMPAS capsule.
-    **kwargs : dict, optional
-        Additional keyword arguments.
-
-    """
-
-    def __init__(self, capsule, **kwargs):
-        super(RhinoCapsuleObject, self).__init__(geometry=capsule, **kwargs)
+    """Scene object for drawing capsule shapes."""
 
     def draw(self):
         """Draw the capsule associated with the scene object.